requirements.txt
Package                Version    
---------------------- -----------
asgiref                3.2.7      
beautifulsoup4         4.9.0      
certifi                2020.4.5.1 
cffi                   1.14.0     
chardet                3.0.4      
cryptography           2.9        
defusedxml             0.6.0      
Django                 2.2.9      
django-allauth         0.41.0     
django-bootstrap4      1.1.1      
django-crispy-forms    1.9.0      
django-redis           4.11.0     
django-social-auth     0.7.28     
django-tinymce4-lite   1.7.5      
djangocodemirror       2.1.0      
evileg-core            0.0.5      
httplib2               0.17.1     
idna                   2.9        
jsmin                  2.2.2      
Markdown               3.2.1      
oauth2                 1.9.0.post1
oauthlib               3.1.0      
Pillow                 7.1.1      
pip                    19.2.3     
pycparser              2.20       
PyJWT                  1.7.1      
python-openid          2.2.5      
python3-openid         3.1.0      
pytz                   2019.3     
redis                  3.4.1      
requests               2.23.0     
requests-oauthlib      1.3.0      
setuptools             41.2.0     
six                    1.14.0     
social-auth-app-django 3.1.0      
social-auth-core       3.3.3      
soupsieve              2.0        
sqlparse               0.3.1      
urllib3                1.25.8     
